# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iPF-Fund-request-real",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "local Server"
        },
        {
            "url": "http://127.0.0.1:8000/",
            "description": "local Server"
        }
    ],
    "paths": {
...
        "/api/Admin/manage-project": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "fetch projects",
                "operationId": "fde5d20f3cae2dd88da5b69b9cb34cea",
                "responses": {
                    "200": {
                        "description": " list of projects",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "create project",
                "operationId": "e8f9090772bc68628494c4251e473e13",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project": {
                                        "description": "project",
                                        "format": "string",
                                        "default": "Transport"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "project created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "created successfully"
                                        },
                                        "project": {
                                            "properties": {
                                                "id": {
                                                    "description": "id",
                                                    "format": "string",
                                                    "default": "ovanmdokvnmo-vakv-vks-vka"
                                                },
                                                "project": {
                                                    "description": "project",
                                                    "format": "string",
                                                    "default": "Transport"
                                                },
                                                "created_at": {
                                                    "description": "created_at",
                                                    "format": "string",
                                                    "default": "2022-08-30T06:55:08.000000Z"
                                                },
                                                "updated_at": {
                                                    "description": "updated_at",
                                                    "format": "string",
                                                    "default": "2022-08-30T06:55:08.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fields required",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/Admin/manage-project/{project_id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "update project",
                "operationId": "bd8bcc346c424f8cab06eb91a84407b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project": {
                                        "description": "project",
                                        "format": "string",
                                        "default": "Transport"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "user updated successfully"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "description": "id",
                                                    "format": "string",
                                                    "default": "kdmvksdm-4324kn-fdv"
                                                },
                                                "project": {
                                                    "description": "project",
                                                    "format": "string",
                                                    "default": "Transport"
                                                },
                                                "created_at": {
                                                    "description": "created_at",
                                                    "format": "string",
                                                    "default": "2022-08-30T06:55:08.000000Z"
                                                },
                                                "updated_at": {
                                                    "description": "updated_at",
                                                    "format": "string",
                                                    "default": "2022-08-30T06:55:08.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fields required",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "delete-project",
                "operationId": "7d418e725b709aa98acf3972db3799b8",
                "responses": {
                    "200": {
                        "description": " delete project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "project deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "code",
                                            "format": "string",
                                            "default": "404"
                                        },
                                        "error": {
                                            "description": "error",
                                            "format": "string",
                                            "default": "project not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}